chore: fix all clippy warnings — zero warnings across workspace
- Replace Default::default()+field assignment with struct literals (market.rs, parameters.rs, sim.rs)
- Replace wildcard _ => {} with explicit ChallengeResult::None => {}
- Add #[expect(clippy::needless_range_loop)] for intentional parallel-array indexing (4 loops)
- Add #[expect(clippy::struct_field_names)] for CognitiveBudgetTracker
- Auto-fix: .is_multiple_of(), manual assign ops, redundant closures, map_or simplifications
- All 600 tests pass, zero clippy warnings

+++ crates/mindstrata-sim/src/social/relationship_stages.rs
@@ -119,7 +119,7 @@ pub fn try_advance_stage(
 ) -> Option<RelationshipStage> {
     let next = current.next_positive()?;
 
-    if interactions as u32 >= min_interactions_for_stage(next)
+    if interactions >= min_interactions_for_stage(next)
         && trust >= trust_threshold_for_stage(next)
         && affection >= affection_threshold_for_stage(next)
     {
